UserService 인터페이스 도입
