+++ src/main/java/springbook/user/service/UserServiceImpl.java
@@ -1,0 +1,81 @@
+package springbook.user.service;
+
+import java.util.List;
+import org.springframework.mail.MailSender;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSenderImpl;
+import springbook.user.dao.UserDao;
+import springbook.user.domain.Level;
+import springbook.user.domain.User;
+
+
+// UserService Interface 를 구현한 클래스는 두개다.
+// Impl에서는 Transaction 을 제외한 나머지 로직을 담당하자
+public class UserServiceImpl implements UserService {
+
+    UserDao userDao;
+    private MailSender mailSender;
+
+    public void setUserDao(UserDao userDao) {
+        this.userDao = userDao;
+    }
+
+    public void setMailSender(MailSender mailSender) {
+        this.mailSender = mailSender;
+    }
+
+    private void upgradeLevels() {
+        List<User> users = userDao.getAll();
+        for (User user : users) {
+            if (canUpgradeLevel(user)) {
+                upgradeLevel(user);
+            }
+        }
+        // 비지니스 로직 끝
+    }
+
+    protected void upgradeLevel(User user) {
+        if (user.getLevel() == Level.BASIC) {
+            user.setLevel(Level.SILVER);
+        } else if (user.getLevel() == Level.SILVER) {
+            user.setLevel(Level.GOLD);
+        }
+    }
+
+    private boolean canUpgradeLevel(User user) {
+        Level currentLevel = user.getLevel();
+        switch (currentLevel) {
+            case BASIC:
+                return (user.getLogin() >= 50);
+            case SILVER:
+                return (user.getRecommend() >= 30);
+            case GOLD:
+                return false;
+            default:
+                throw new IllegalArgumentException("unknown level:" + currentLevel);
+        }
+    }
+
+    public void add(User user) {
+        if (user.getLevel() == null) {
+            user.setLevel(Level.BASIC);
+        }
+        userDao.add(user);
+    }
+
+    private void sendUpgradeEMail(User user) {
+        JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
+        mailSender.setHost("mail.server.com");
+
+        SimpleMailMessage mailMessage = new SimpleMailMessage();
+        mailMessage.setTo(user.getEmail());
+        mailMessage.setFrom("useradmin@ksug.org");
+        mailMessage.setSubject("Upgrade 안내");
+        mailMessage.setText("사용자의 등급이 " + user.getLevel());
+
+        mailSender.send(mailMessage);
+
+    }
+
+
+}
