+++ src/main/java/springbook/user/service/UserServiceTx.java
@@ -1,0 +1,45 @@
+package springbook.user.service;
+
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.support.DefaultTransactionDefinition;
+import springbook.user.domain.User;
+
+// UserService Interface 를 구현한 클래스는 두개다.
+// Tx에서는 Transaction 관련된 부분만 다루자
+public class UserServiceTx implements UserService {
+
+    private UserService userService;
+
+
+    PlatformTransactionManager transactionManager;
+
+    public void setUserService(UserService userService) {
+        this.userService = userService;
+    }
+
+    public void setTransactionManager(PlatformTransactionManager transactionManager) {
+        this.transactionManager = transactionManager;
+    }
+
+
+    @Override
+    public void add(User user) {
+        this.userService.add(user);
+    }
+
+    @Override
+    public void upgradeLevels() {
+        TransactionStatus status = this.transactionManager
+            .getTransaction(new DefaultTransactionDefinition());
+        try {
+            userService.upgradeLevels();
+
+            this.transactionManager.commit(status);
+        } catch (RuntimeException e) {
+            this.transactionManager.rollback(status);
+            throw e;
+        }
+    }
+
+}
