+++ src/main/resources/test-applicationContext.xml
@@ -3,9 +3,11 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="userService" class="springbook.user.service.UserService">
-        <property name="userDao" ref="userDao"/>
+    <bean id="userServiceTx" class="springbook.user.service.UserServiceTx">
         <property name="transactionManager" ref="transactionManager"/>
+    </bean>
+    <bean id="userService" class="springbook.user.service.UserServiceImpl">
+        <property name="userDao" ref="userDao"/>
         <property name="mailSender" ref="mailSender"/>
     </bean>
     <bean id="mailSender" class="springbook.user.service.DummyMailSender"/>
